Revert "Added src directory and moved app and logic to src"
This reverts commit 01ac74d04a3f42e599beba04167f6fed62291504.
