+++ logic/logic1.py
